requirements.txt
biopython
matplotlib
numpy
bcbio-gff
colorlog
pandas